+++ ApiBcosHealth/src/cl/bcos/option/ApiUpdateEnfermedadesCronicas.java
@@ -43,7 +43,7 @@ public class ApiUpdateEnfermedadesCronicas extends ServerResource {
         Log.debug(Thread.currentThread().getStackTrace()[1].getMethodName());
         String path = getRequest().getResourceRef().getHostIdentifier() + getRequest().getResourceRef().getPath();
         Log.info("path : " + path);
-        
+
         Status status = null;
         String message = "ok";
         Gson gson = new GsonBuilder().disableHtmlEscaping().setPrettyPrinting().create();
@@ -74,44 +74,49 @@ public class ApiUpdateEnfermedadesCronicas extends ServerResource {
                     String apellido_usuario = jwt.getJwt().getValue("LastName").toString();
                     String nombre_completo = nombre_usuario + " " + apellido_usuario;
                     String empresa = jwt.getJwt().getValue("empresaName").toString();
+                    String roles = jwt.getJwt().getValue("Roles").toString();
 
                     Log.info(usuario_creador);
+                    if (roles.contains("SUPER-ADMIN") || roles.contains("MEDICO") || roles.contains("ADMIN") || roles.contains("RECEPCION")) {
+                        if (accion.equalsIgnoreCase(DELETE_CRONICA)) {
+                            if (LFEnfermedadesCronicas.deleteEnfermedadCronica(CronicaRowId, empresa) == 1) {
 
-                    if (accion.equalsIgnoreCase(DELETE_CRONICA)) {
-                        if (LFEnfermedadesCronicas.deleteEnfermedadCronica(CronicaRowId, empresa) == 1) {
+                                Log.info("DELETE_OK");
+                                status = Status.SUCCESS_OK;
+                                message = "UPDATE_OK";
 
-                            Log.info("DELETE_OK");
-                            status = Status.SUCCESS_OK;
-                            message = "UPDATE_OK";
+                            } else {
+                                Log.info("Error de Udpate/Delete");
+                                message = "UPDATE_NO_OK";
+                                status = Status.CLIENT_ERROR_BAD_REQUEST;
+                            }
 
+                        } else if (accion.equalsIgnoreCase(UPDATE_CRONICA)) {
+
+                            if (LFEnfermedadesCronicas.updateEnfermedadesCronicas(CronicaRowId,
+                                    cronica_c_name, cronica_c_obs,
+                                    usuario_creador, nombre_completo, empresa) == 1) {
+
+                                Log.info("UPDATE OK");
+                                status = Status.SUCCESS_OK;
+                                message = "UPDATE_OK";
+                            } else {
+                                Log.info("Error de Udpate/Delete");
+                                message = "UPDATE_NO_OK";
+                                status = Status.CLIENT_ERROR_BAD_REQUEST;
+                            }
                         } else {
+
                             Log.info("Error de Udpate/Delete");
                             message = "UPDATE_NO_OK";
                             status = Status.CLIENT_ERROR_BAD_REQUEST;
-                        }
 
-                    } else if (accion.equalsIgnoreCase(UPDATE_CRONICA)) {
-
-                        if (LFEnfermedadesCronicas.updateEnfermedadesCronicas(CronicaRowId,
-                                cronica_c_name, cronica_c_obs,
-                                usuario_creador, nombre_completo, empresa) == 1) {
-
-                            Log.info("UPDATE OK");
-                            status = Status.SUCCESS_OK;
-                            message = "UPDATE_OK";
-                        } else {
-                            Log.info("Error de Udpate/Delete");
-                            message = "UPDATE_NO_OK";
-                            status = Status.CLIENT_ERROR_BAD_REQUEST;
                         }
                     } else {
-
-                        Log.info("Error de Udpate/Delete");
-                        message = "UPDATE_NO_OK";
-                        status = Status.CLIENT_ERROR_BAD_REQUEST;
-
+                        status = Status.CLIENT_ERROR_UNAUTHORIZED;
+                        Log.error("Perfil sin acceso");
+                        message = ERROR_TOKEN;
                     }
-
                 } catch (Exception e) {
                     Log.error(e.toString());
                     status = Status.CLIENT_ERROR_BAD_REQUEST;
